review name changes, list of ips to ipnet
