+++ Makefile
@@ -3,7 +3,7 @@ all: zgrab2
 .PHONY: clean zgrab2
 
 zgrab2: 
-	cd main && go build -o zgrab2 && ./zgrab2 multiple -c mult.ini 
+	cd main && go build -o zgrab2  
 
 clean:
 	go clean
