+++ utility.go
@@ -14,10 +14,12 @@ func init() {
 	parser = flags.NewParser(&config, flags.Default)
 }
 
-func AddCommand(command string, shortDescription string, longDescription string, m Module) (*flags.Command, error) {
+// AddCommand adds a module to the parser and returns a pointer to a flags.command object or an error
+func AddCommand(command string, shortDescription string, longDescription string, m ScanModule) (*flags.Command, error) {
 	return parser.AddCommand(command, shortDescription, longDescription, m)
 }
 
+// ParseFlags abstracts away the parser and validates the framework configuration (global options) immediately after parsing
 func ParseFlags() ([]string, error) {
 	r, err := parser.Parse()
 	if err == nil {
@@ -26,8 +28,8 @@ func ParseFlags() ([]string, error) {
 	return r, err
 }
 
-// ParseInput takes input and parses it into either a list of IP addresses, domain name, or errors
-func ParseInput(s string) ([]net.IP, string, error) {
+// ParseInput takes input as a string and parses it into either an IPNet (may have empty mask and just contain IP) , domain name, or errors, may return both IPNet and domain name
+func ParseInput(s string) (*net.IPNet, string, error) {
 	i := strings.IndexByte(s, ',')
 	j := strings.IndexByte(s, '/')
 
@@ -35,27 +37,21 @@ func ParseInput(s string) ([]net.IP, string, error) {
 	case i == -1 && j == -1:
 		//just ip or domain
 		if ip := net.ParseIP(s); ip != nil {
-			return []net.IP{ip}, "", nil
+			return &net.IPNet{IP: ip}, "", nil
 		} else {
 			ips, err := net.LookupIP(s)
 			if err != nil {
 				return nil, "", err
 			}
-			return ips, s, nil
+			return &net.IPNet{IP: ips[0]}, s, nil //only return first IP after a lookup
 		}
 	case i == -1:
 		//cidr block
-		ip, ipnet, err := net.ParseCIDR(s)
+		_, ipnet, err := net.ParseCIDR(s)
 		if err != nil {
 			return nil, "", err
 		}
-		var ips []net.IP
-		for ip := ip.Mask(ipnet.Mask); ipnet.Contains(ip); incrementIP(ip) {
-			ips = append(ips, duplicateIP(ip))
-		}
-
-		return ips, "", nil
-
+		return ipnet, "", nil
 	case j == -1:
 		//ip,domain
 		str := strings.Split(s, ",")
@@ -63,7 +59,7 @@ func ParseInput(s string) ([]net.IP, string, error) {
 			return nil, "", errors.New("malformed input")
 		}
 		if ip := net.ParseIP(str[0]); ip != nil {
-			return []net.IP{ip}, str[1], nil
+			return &net.IPNet{IP: ip}, str[1], nil
 		}
 		return nil, str[1], nil
 	}
